# `PQ::Connection#send_bind_message`: write entire slice to socket at once